Support build on Windows
Prerequisite: curl

On Windows, the source code amalgamation is donwloaded from SQLite web
page with curl command.

Why not build from the repository source code? It's because in order to
create the amalgamation on Windows, we need Tcl command. To avoid the
dependencies on Tcl, I choose to download the amalgamation from the web.

The current limitation of this approach is the version of sqlite is
hard-coded in the cmake script. So if the SQLite's version is
incremented and the URL gets unavaialble, this script fails.